firmware-imx: bump to version 5.2
Use a different versioning schema based on package versioning
instead of BSP version. This way, if a new BSP version does
not bump this package, we can keep this package unchanged as well.

VPU firmware upgraded to v3.1.1_r46067
- Low Latency fixes for H.264 encoder
- Wrong frame number fix for H.264 decoder
- Decode failure fix for H.264 decoder with multi-SPS/PPS headers

The package adds the EPDC waveforms and the needed changes to properly
install those were done in the firmware-imx.mk file.

This patch is based on the Yocto equivalent:
https://github.com/Freescale/meta-fsl-arm/commit/d869826
https://github.com/Freescale/meta-fsl-arm/commit/075a5b1

This package has been implicitely tested through gstreamer as the
plugins rely on them for vpu decoding for instance:
 # gst-launch-1.0 filesrc location=/root/tears_of_steel_1080p.webm ! \
   matroskademux ! imxvpudec ! imxipuvideosink

Signed-off-by: Gary Bisson <gary.bisson@boundarydevices.com>
[Thomas:
 - fix typo in commit logs
 - remove ; \ and ; at the end of commands, since they were not
   needed.]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>