protobuf: introduce BR2_PACKAGE_PROTOBUF_ARCH_SUPPORTS
The architecture dependencies of protobuf are going to change, and
they are already duplicated between protobuf, ola and mosh. In order
to factorize the expression of those dependencies, this commit
introduces BR2_PACKAGE_PROTOBUF_ARCH_SUPPORTS.

Note that we include in this hidden Config.in option both the target
architecture dependency and the host architecture dependency.

Finally, this commit also fixes a real mistake in the mosh Config.in
file, where the condition on the architecture dependency for the
Config.in comment was inverted: we only want to show the Config.in
comment when we are on supported architectures.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>