Config.in.legacy: fix separator for 2016.05 legacy options
Since we opened up the next branch, a number of options have been
added to Config.in.legacy. However, they are located under the "legacy
options removed in 2016.02", which is not correct. This commit fixes
that by adding the "legacy options removed in 2016.05" header, and
restoring the one related to 2016.02 at the right location.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>