package/jack2: add comment about celt and eigen
The optinal dependency on celt is not added, because we only have
celt-0.5.1 and jack2 requires celt >= 0.5.2 (even though it does look
for celt >= 0.5.0, it does not build with celt- < 0.5.2). Since we
cannot upgrade celt (for now), we just never make jack2 depend on
celt051, and it won't find it either, as it just searches for celt.pc
and we have celt051.pc.

As well, the dependency on eigen is only useful in conjunction with
gtkiostream, for which we do not have a package. So, we don't need to
depend on eigen.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>